+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.14.0-rc.44",
-  genVersion: "2.727.9",
-  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.44 2.727.9 2.0 @orq-ai/node",
+  sdkVersion: "3.14.0-rc.45",
+  genVersion: "2.728.0",
+  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.45 2.728.0 2.0 @orq-ai/node",
 } as const;