+++ packages/orq-rc/RELEASES.md
@@ -4439,3 +4439,13 @@ Based on:
 - [typescript v3.14.0-rc.44] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.44 - packages/orq-rc
+
+## 2025-10-22 05:17:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.638.0 (2.728.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.45 - packages/orq-rc